refactor(demo): extract useDemoState hook, HeroIdentifier panel, layout primitives
App.tsx drops from 442 lines to 106 — pure composition root. Every piece
it used to own moves to a named home:

  demo/src/hooks/useDemoState.ts        — the single ownership boundary for
                                          input, paletteKey, metric, URL
                                          sync, derived identify/resolve/
                                          convert values, stage gels,
                                          applyPreset action
  demo/src/data/warhammer.ts            — BYO palette + per-faction
                                          invocation map
  demo/src/data/presets.ts              — hero-section preset list
  demo/src/components/DemoPanel.tsx     — layout primitives (ActHeader,
                                          SubChapterHeader, DemoPanel);
                                          the bordered-below-a-black-bar
                                          pattern used everywhere
  demo/src/components/Masthead.tsx      — page header (eyebrow, Wordmark,
                                          tagline, github/npm links)
  demo/src/components/Footer.tsx        — colophon + Bauhaus Modern
                                          attribution + Pantone notice +
                                          the Simiancraft house credit
  demo/src/components/HeroIdentifier.tsx — the scrub-a-color panel (picker,
                                          PaletteTiles, metric selector,
                                          presets row, result grid, and
                                          its own LiveSnippet)

Design decisions captured here:

1. useState, not useReducer — for three primitives plus a preset action,
   the reducer boilerplate would outweigh the wins. `applyPreset` is a
   named action returned from the hook so if the vocabulary grows, the
   internals can flip to useReducer without changing callers.

2. Translator and Eyedropper keep their *local* state (side tracking,
   file / webcam / pinned, etc.). Promoting them to global via the hook
   would couple them to App.tsx and kill reusability. Shared state is
   exactly the three things that every panel legitimately reads from.

3. HeroIdentifier is a dumb presentational component — no hooks of its
   own. It receives exactly what it renders from props. Easier to test,
   easier to storyboard, and the data flow from hook → prop → JSX is
   directly legible at the composition root.

4. DemoPanel owns only the chrome (eyebrow/title/kicker header + outer
   border). Content components own their own padding + background fill,
   so the hero's paper tint, the translator's paper tint, and
   KandinskyBYO's grid layout all compose cleanly without DemoPanel
   having to know about any of them.

Everything still green: typecheck, biome, 381 tests, knip, vite build.
Manual verification: preset click fires applyPreset → URL query string
updates to ?c=#1DB954&cs=pantone&m=deltaE2000, gels tint, all downstream
panels react — behavior identical to pre-refactor.
